requirements.txt
# Core data processing
pandas==2.1.3
numpy==1.25.2

# Technical analysis
ta-lib==0.4.28
ta==0.10.2

# Market data APIs
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0

# Machine Learning / AI
scikit-learn==1.3.2
tensorflow==2.15.0

# Async programming
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Configuration management
python-dotenv==1.0.0
pydantic==2.5.0

# Notifications
slack-sdk==3.26.1
twilio==8.12.0

# Database
SQLAlchemy==2.0.23
sqlite3

# Logging and monitoring
structlog==23.2.0
rich==13.7.0

# API clients (placeholders for future integration)
# etrade-python-client  # Will need custom implementation
# crypto-com-exchange   # Will need API integration

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1